added checks

+++ app.py
@@ -1,8 +1,7 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request, abort, jsonify
 from flask_sqlalchemy import SQLAlchemy 
 from flask_marshmallow import Marshmallow 
 import os
-import datetime
 
 # Init app
 app = Flask(__name__)
@@ -21,18 +20,16 @@ class Meme(db.Model):
   name = db.Column(db.String(100))
   url = db.Column(db.String(500))
   caption = db.Column(db.String(500))
-  date = db.Column(db.String(500))
 
-  def __init__(self, name, url, caption, date):
+  def __init__(self, name, url, caption):
     self.name = name
     self.url = url
     self.caption = caption
-    self.date = date
 
 # Meme Schema
 class MemeSchema(ma.Schema):
   class Meta:
-    fields = ('id', 'name', 'url', 'caption', 'date')
+    fields = ('id', 'name', 'url', 'caption')
 
 # Init schema
 meme_schema = MemeSchema()
@@ -41,18 +38,39 @@ memes_schema = MemeSchema(many=True)
 # Create a Meme
 @app.route('/memes', methods=['POST'])
 def add_meme():
+  if not "name" in request.json:
+    result = {
+      "error": "name is required"
+    }
+    return jsonify(result), 400
+  
+  if not "url" in request.json:
+    result = {
+      "error": "url is required"
+    }
+    return jsonify(result), 400
+  
+  if not "caption" in request.json:
+    result = {
+      "error": "caption is required"
+    }
+    return jsonify(result), 400
+  
   name = request.json['name']
   url = request.json['url']
   caption = request.json['caption']
-  date = datetime.datetime.now().strftime("%b-%d-%Y, %H:%M")
-  print(date)
 
-  new_meme = Meme(name, url, caption, date)
+  new_meme = Meme(name, url, caption)
 
   db.session.add(new_meme)
   db.session.commit()
 
-  return meme_schema.jsonify(new_meme)
+  result = meme_schema.dump(new_meme)
+  del result['name']
+  del result['url']
+  del result['caption']
+
+  return jsonify(result)
 
 # Get All Memes
 @app.route('/memes', methods=['GET'])
@@ -69,6 +87,11 @@ def get_memes():
 @app.route('/memes/<id>', methods=['GET'])
 def get_meme(id):
   meme = Meme.query.get(id)
+  if meme is None:
+    result = {
+      "error": "Id not found"
+    }
+    return jsonify(result), 404
   return meme_schema.jsonify(meme)
 
 # Update a Meme
@@ -76,13 +99,18 @@ def get_meme(id):
 def update_meme(id):
   meme = Meme.query.get(id)
 
-  name = request.json['name']
-  url = request.json['url']
-  caption = request.json['caption']
+  if meme is None:
+    result = {
+      "error": "Id not found"
+    }
+    return jsonify(result), 404
 
-  meme.name = name
-  meme.url = url
-  meme.caption = caption
+  if "name" in request.json:
+    meme.name = request.json['name']
+  if "url" in request.json:
+    meme.url = request.json['url']
+  if "caption" in request.json:
+    meme.caption = request.json['caption']
 
   db.session.commit()
 
@@ -92,6 +120,11 @@ def update_meme(id):
 @app.route('/memes/<id>', methods=['DELETE'])
 def delete_meme(id):
   meme = Meme.query.get(id)
+  if meme is None:
+    result = {
+      "error": "Id not found"
+    }
+    return jsonify(result), 404
   db.session.delete(meme)
   db.session.commit()
